feat(controller): refactoring rename internal/service folder to internal/controller
